update rules for services CRUD update auth flow
