+++ middleware/auth.js
@@ -1,0 +1,11 @@
+export default ({ store, error, redirect }) => {
+  if (!store.getters.isLogged) {
+    // redirect('/')
+    // error({
+    //   message: 'You are not connected',
+    //   statusCode: 403
+    // })
+  }
+
+  return true
+}
