+++ pages/app.vue
@@ -108,7 +108,7 @@
               </div>
               <button
                 class="top-0 right-0 m-2 absolute text-gray-100 hover:text-white"
-                v-if="authUser.uid"
+                v-if="service.owner === authUser.uid"
                 @click="edit(service)"
               >
                 <fa class="w-4" :icon="['fa', 'edit']" />
@@ -195,43 +195,6 @@ export default {
         {
           name: "Retirar em mãos",
           status: true
-        }
-      ],
-      items: [
-        {
-          id: 1,
-          title: "Item 1",
-          description:
-            "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptates consequuntur iure consequatur tempora ab impedit sed",
-          image: ""
-        },
-        {
-          id: 2,
-          title: "Item 2",
-          description:
-            "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptates consequuntur iure consequatur tempora ab impedit sed",
-          image: ""
-        },
-        {
-          id: 3,
-          title: "Item 3",
-          description:
-            "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptates consequuntur iure consequatur tempora ab impedit sed",
-          image: ""
-        },
-        {
-          id: 4,
-          title: "Item 4",
-          description:
-            "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptates consequuntur iure consequatur tempora ab impedit sed",
-          image: ""
-        },
-        {
-          id: 5,
-          title: "Item 5",
-          description:
-            "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptates consequuntur iure consequatur tempora ab impedit sed",
-          image: ""
         }
       ]
     };
